Issue #17 Convert character_numbert to UINT16 as pushing by 4 bits caused an overflow -e Original/build commit ID: ae984b8687c7b7969f2f10b4b919985a097d005b

+++ window_text.c
@@ -13,7 +13,7 @@ void show_window_text(UINT8 *text)
     UINT8 itx_x = 3U;
     UINT8 itx_y = 3U;
     UINT8 text_index = 0;
-    UINT8 character_number;
+    UINT16 character_number;
     UBYTE tile_data;
 
     while (1)
